locli:  massive cleanups
