+++ bench/locli/app/locli.hs
@@ -6,8 +6,7 @@ import           Control.Monad.Trans.Except.Exit (orDie)
 import qualified Options.Applicative as Opt
 
 import           Cardano.Analysis.TopHandler
-import           Cardano.Unlog.Parsers (opts, pref)
-import           Cardano.Unlog.Run (renderCommandError, runCommand)
+import           Cardano.Command (opts, pref, renderCommandError, runCommand)
 
 
 main :: IO ()
