+++ bench/locli/src/Cardano/Analysis/BlockProp.hs
@@ -52,6 +52,7 @@ import Data.Distribution
 import Cardano.Analysis.API
 import Cardano.Analysis.Chain
 import Cardano.Analysis.ChainFilter
+import Cardano.Analysis.Ground
 import Cardano.Analysis.Run
 import Cardano.Analysis.Version
 import Cardano.Unlog.LogObject  hiding (Text)
